Add basic validate for user data

+++ src/main/java/com/example/hotrotinhthue/repository/BangNopThueRepo.java
@@ -3,5 +3,5 @@ package com.example.hotrotinhthue.repository;
 import com.example.hotrotinhthue.model.BangNopThue;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BangNopThueRepo extends JpaRepository<BangNopThue, Integer> {
+public interface BangNopThueRepo extends JpaRepository<BangNopThue, Long> {
 }
